Move download_from_repository into a separate file
